Move observer code to avoid expliticit statements
By moving all the methods bodies to the header file it's no longer necessary to explicitly make statements of how the "Subject" class is being used.
