+++ Axle/src/Other/ObserverPattern/Observer.hpp
@@ -16,16 +16,37 @@ namespace Axle {
 	template<typename... Args>
 	class AXLE_API Subscription {
 	public:
-		Subscription(Subject<Args...>* subject, int id);
-		~Subscription();
+		Subscription(Subject<Args...>* subject, int id)
+			: m_subject(subject), m_id(id) { }
+		~Subscription() {
+			Unsubscribe();
+		}
 
-		Subscription(Subscription&& other) noexcept;
-		Subscription& operator=(Subscription&& other) noexcept;
+		Subscription(Subscription&& other) noexcept
+			: m_subject(other.m_subject), m_id(other.m_id), m_unsubscribed(other.m_unsubscribed) {
+			other.m_subject = nullptr;
+		}
+		Subscription& operator=(Subscription&& other) noexcept {
+			if (this != &other) {
+				Unsubscribe();
+
+				m_subject = other.m_subject;
+				m_id = other.m_id;
+				m_unsubscribed = other.m_unsubscribed;
+
+				other.m_subject = nullptr;
+			}
+		}
 
 		Subscription(const Subscription&) = delete;
 		Subscription& operator=(const Subscription&) = delete;
 
-		void Unsubscribe();
+		void Unsubscribe() {
+			if (!m_unsubscribed && m_subject) {
+				m_subject->Unsubscribe(m_id);
+				m_unsubscribed = true;
+			}
+		}
 
 	private:
 		Subject<Args...>* m_subject;
@@ -39,15 +60,25 @@ namespace Axle {
 	public:
 		using HandlerType = std::function<void(Args...)>;
 
-		virtual Subscription<Args...> Subscribe(const HandlerType& handler);
+		virtual Subscription<Args...> Subscribe(const HandlerType& handler) {
+			int id = m_nextId++;
+			m_handlers[id] = handler;
+			return Subscription<Args...>(this, id);
+		}
 
 		virtual ~Subject() = default;
-		virtual void Notify(Args... args);
+		virtual void Notify(Args... args) {
+			for (auto& [id, handler] : m_handlers) {
+				handler(args...);
+			}
+		}
 
 	protected:
 		friend class Subscription<Args...>;
 
-		virtual void Unsubscribe(int id);
+		virtual void Unsubscribe(int id) {
+			m_handlers.erase(id);
+		}
 
 		std::unordered_map<int, HandlerType> m_handlers;
 		int m_nextId = 0;
